Deleting and Putting Search Terms First

+++ lib/search/infrastructure/search_history_repository.dart
@@ -25,16 +25,48 @@ class SearchHistoryRepository {
         .map((records) => records.reversed.map((e) => e.value).toList());
   }
 
-  Future<void> addSearchTerm(String term) async {
-    ['flutter', 'dart', 'github', 'riverpod'].reversed;
+  Future<void> addSearchTerm(String term) =>
+      _addSearchTerm(term, _sembastDatabase.instance);
 
-    await _store.add(_sembastDatabase.instance, term);
-    final count = await _store.count(_sembastDatabase.instance);
+  Future<void> deleteSearchTerm(String term) =>
+      _deleteSearchTerm(term, _sembastDatabase.instance);
+
+  Future<void> putSearchTermFirst(String term) async {
+    await _sembastDatabase.instance.transaction((transaction) async {
+      await _deleteSearchTerm(term, transaction);
+      await _addSearchTerm(term, transaction);
+    });
+  }
+
+  Future<void> _addSearchTerm(String term, DatabaseClient dbClient) async {
+    final existingKey = await _store.findKey(
+      dbClient,
+      finder: Finder(
+        filter: Filter.custom((record) => record.value == term),
+      ),
+    );
+
+    if (existingKey != null) {
+      putSearchTermFirst(term);
+      return;
+    }
+
+    await _store.add(dbClient, term);
+    final count = await _store.count(dbClient);
     if (count > historyLength) {
       await _store.delete(
-        _sembastDatabase.instance,
+        dbClient,
         finder: Finder(limit: count - historyLength),
       );
     }
+  }
+
+  Future<void> _deleteSearchTerm(String term, DatabaseClient dbClient) async {
+    await _store.delete(
+      dbClient,
+      finder: Finder(
+        filter: Filter.custom((record) => record.value == term),
+      ),
+    );
   }
 }
